Move simple workflow into workflow module
